model/MM_Det/requirements.txt
torch==2.2.1
torchvision==0.17.1
transformers==4.37.2
tokenizers==0.15.1
sentencepiece==0.1.99
shortuuid==1.0.13
accelerate==0.21.0
bitsandbytes==0.43.1
opencv-python==4.9.0.80
peft==0.10.0
pydantic==2.7.1
markdown2[all]
numpy==1.26.3
pillow==10.2.0
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.13.0
einops==0.6.1
einops-exts==0.0.4
tensorboardX==2.6.2.2
timm==0.9.16
fastapi
uvicorn
python-multipart